Clean up stuff and fix show_bytes
