+++ platform/brains2/src/demos/flexio-leds/CMakeLists.txt
@@ -1,12 +1,12 @@
 CMAKE_MINIMUM_REQUIRED (VERSION 3.10.0)
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
-project(rust-test)
-set(EXECUTABLE_NAME rust-test.elf)
+project(flexio-leds)
+set(EXECUTABLE_NAME flexio-leds.elf)
 
 
 add_executable(${EXECUTABLE_NAME}
-    ${CMAKE_CURRENT_SOURCE_DIR}/rust_test.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/flexio_leds.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/fsl_flexio.c
 )
